# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: StudyRover API
  version: 0.1.0
  description: StudyRover Phase 1 — the tutoring core. A parent sets up subjects/topics/questions/exams;
    a student takes exams; the system produces scores + mastery and emits a ScoreEvent. JSON over HTTP;
    parent cookie session required except for the student exam + student-auth endpoints.
  license:
    name: Proprietary
servers:
- url: /api
  description: StudyRover API, served by the Go binary under the /api base path.
security:
- cookieSession: []
...
paths:
  /subjects:
    get:
      operationId: listSubjects
      tags:
      - subjects
      summary: List subjects
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A page of subjects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfSubject'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createSubject
      tags:
      - subjects
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubject'
      responses:
        '201':
          description: The created subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        default:
          $ref: '#/components/responses/Problem'
  /subjects/{id}:
    parameters:
    - $ref: '#/components/parameters/IdPath'
    get:
      operationId: getSubject
      tags:
      - subjects
      summary: Get a subject
      responses:
        '200':
          description: The subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateSubject
      tags:
      - subjects
      summary: Update a subject (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subject'
      responses:
        '200':
          description: The updated subject.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteSubject
      tags:
      - subjects
      summary: Delete a subject
      responses:
        '204':
          description: The subject was deleted.
        default:
          $ref: '#/components/responses/Problem'
...
  /exam-definitions/{id}:
    parameters:
    - $ref: '#/components/parameters/IdPath'
    get:
      operationId: getExamDefinition
      tags:
      - examDefinitions
      summary: Get an exam definition
      responses:
        '200':
          description: The exam definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamDefinition'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateExamDefinition
      tags:
      - examDefinitions
      summary: Update an exam definition (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamDefinition'
      responses:
        '200':
          description: The updated exam definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamDefinition'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteExamDefinition
      tags:
      - examDefinitions
      summary: Delete an exam definition
      responses:
        '204':
          description: The exam definition was deleted.
        default:
          $ref: '#/components/responses/Problem'
...
  /attempts:
    get:
      operationId: listAttempts
      tags:
      - attempts
      summary: List exam attempts for a student
      parameters:
      - name: studentId
        in: query
        required: false
        schema:
          type: string
        description: Filter attempts to a single student.
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A page of exam attempts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfExamAttempt'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: startAttempt
      tags:
      - attempts
      summary: Start a new exam attempt
      description: Assembles the exam and persists an in_progress attempt, returning the attempt plus
        the delivered questions (without the answer key).
      security:
      - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAttempt'
      responses:
        '201':
          description: The started attempt and its delivered questions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartAttemptResult'
        default:
          $ref: '#/components/responses/Problem'
...
  /auth/register:
    post:
      operationId: authRegister
      tags:
      - auth
      summary: Parent WebAuthn registration ceremony (begin/finish)
      description: Drives the parent WebAuthn registration ceremony. The opaque ceremony payload (begin
        options or the authenticator response) is passed through to go-webauthn. On finish, a Session
        is returned and the session cookie is set.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnCeremony'
      responses:
        '200':
          description: Ceremony options, or a Session on a successful finish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        default:
          $ref: '#/components/responses/Problem'
  /auth/login:
    post:
      operationId: authLogin
      tags:
      - auth
      summary: Parent WebAuthn login ceremony (begin/finish)
      description: Drives the parent WebAuthn authentication ceremony. The opaque ceremony payload is
        passed through to go-webauthn. On finish, a Session is returned and the session cookie is set.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnCeremony'
      responses:
        '200':
          description: Ceremony options, or a Session on a successful finish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        default:
          $ref: '#/components/responses/Problem'
  /auth/student:
    post:
      operationId: authStudent
      tags:
      - auth
      summary: Student sign-in
      description: Signs a student in (optionally with a PIN) and sets the session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentSignIn'
      responses:
        '200':
          description: The student session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: sr_session
      description: HTTP-only session cookie issued by /auth/* after a successful WebAuthn (parent) or
        PIN (student) ceremony. Carries the role + subject id.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource identifier (UUID/cuid).
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      description: 1-based page number to return.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Maximum number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: An error occurred; the body is an RFC 7807-style problem.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    AttemptStatus:
      type: string
      description: Lifecycle state of an exam attempt.
      enum:
      - in_progress
      - submitted
...
    ExamAttempt:
      type: object
      description: A student's attempt at an exam definition. Created in_progress with the assembled questionIds;
        graded fields (scorePct, passed, perTopic, cooldownUntil, submittedAt) are populated after submission.
      additionalProperties: false
      required:
      - id
      - examDefinitionId
      - studentId
      - status
      - questionIds
      - answers
      - startedAt
      properties:
        id:
          type: string
          description: Unique attempt id.
        examDefinitionId:
          type: string
          description: The exam definition this attempt is for.
        studentId:
          type: string
          description: The student taking the exam.
        status:
          allOf:
          - $ref: '#/components/schemas/AttemptStatus'
          description: Current lifecycle state.
          default: in_progress
        questionIds:
          type: array
          description: The assembled question ids, in delivery order.
          items:
            type: string
        answers:
          type: array
          description: The student's answers; empty until questions are answered.
          default: []
          items:
            $ref: '#/components/schemas/Answer'
        scorePct:
          type: integer
          description: Overall score percentage (0–100); present once graded.
          minimum: 0
          maximum: 100
        passed:
          type: boolean
          description: Whether the attempt met the exam pass bar; present once graded.
        perTopic:
          type: array
          description: Per-topic breakdown; present once graded.
          items:
            $ref: '#/components/schemas/PerTopicScore'
        cooldownUntil:
          type: string
          format: date-time
          description: RFC 3339 timestamp until which a new attempt is blocked after a failed attempt;
            absent when there is no active cooldown.
        startedAt:
          type: string
          format: date-time
          description: When the attempt was started.
        submittedAt:
          type: string
          format: date-time
          description: When the attempt was submitted; absent while in_progress.
    StartAttempt:
      type: object
      description: Request body to start a new exam attempt.
      additionalProperties: false
      required:
      - examDefinitionId
      - studentId
      properties:
        examDefinitionId:
          type: string
          description: The exam definition to attempt.
        studentId:
          type: string
          description: The student starting the attempt.
...
    PageOfExamAttempt:
      type: object
      description: A paginated page of exam attempts.
      additionalProperties: false
      required:
      - items
      - total
      - page
      - pageSize
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExamAttempt'
        total:
          type: integer
          description: Total number of attempts matching the query.
          minimum: 0
        page:
          type: integer
          description: 1-based page number.
          minimum: 1
        pageSize:
          type: integer
          description: Number of items per page.
          minimum: 1
...
    Session:
      type: object
      description: The active session identity returned after a successful auth ceremony. Mirrors what
        the server stores behind the sr_session cookie.
      additionalProperties: false
      required:
      - role
      - id
      properties:
        role:
          type: string
          description: Whether the session belongs to a parent or a student.
          enum:
          - parent
          - student
        id:
          type: string
          description: Id of the authenticated parent or student.
    WebAuthnCeremony:
      type: object
      description: Free-form WebAuthn ceremony payload (registration/authentication options or the authenticator's
        response). Opaque — the structure is owned by the WebAuthn libraries, not this contract.
      additionalProperties: true
    Code:
      type: string
      description: 'Machine-readable error category. Maps to the matching HTTP status: VALIDATION->422/400,
        NOT_FOUND->404, UNAUTHORIZED->401, CONFLICT->409, INTERNAL->500.'
      enum:
      - VALIDATION
      - NOT_FOUND
      - UNAUTHORIZED
      - CONFLICT
      - INTERNAL
    Problem:
      type: object
      description: RFC 7807-style problem detail returned for any error response. The HTTP status of the
        response matches the `status` field and the `code` category.
      additionalProperties: false
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
          description: A URI reference identifying the problem type. May be the literal "about:blank"
            when no specific type is defined.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          description: The HTTP status code for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        code:
          $ref: '#/components/schemas/Code'
    ExamType:
      type: string
      description: Kind of exam. `gate` exams are the everyday internet-time gate; `formal` exams are
        heavier graded assessments.
      enum:
      - gate
      - formal
    RewardStyle:
      type: string
      description: How earned internet time relates to the score. `flat` grants the full time for the
        exam size on a pass; `scaled` multiplies by the score%.
      enum:
      - flat
      - scaled
    ExamDefinition:
      type: object
      description: A parent-configured exam template for a subject.
      additionalProperties: false
      required:
      - id
      - subjectId
      - name
      - type
      - scopeTopicIds
      - size
      - passBar
      - cooldownMin
      - rewardStyle
      - createdAt
      properties:
        id:
          type: string
          description: Stable identifier (UUID/cuid).
        subjectId:
          type: string
          description: The subject this exam template belongs to.
        name:
          type: string
          description: Human-readable name for the exam template.
        type:
          $ref: '#/components/schemas/ExamType'
          default: gate
        scopeTopicIds:
          type: array
          description: Topic ids the exam draws from. An empty array means the whole subject.
          items:
            type: string
          default: []
        size:
          type: integer
          description: Number of questions in the exam.
          minimum: 1
          default: 20
        passBar:
          type: integer
          description: Minimum score percentage required to pass.
          minimum: 0
          maximum: 100
          default: 70
        cooldownMin:
          type: integer
          description: Minutes a student must wait after a failed attempt.
          minimum: 0
          default: 10
        rewardStyle:
          $ref: '#/components/schemas/RewardStyle'
          default: flat
        createdAt:
          type: string
          format: date-time
          description: When the exam template was created (RFC 3339).
...
    Subject:
      type: object
      description: A unit of study a parent configures (e.g. "Algebra I").
      required:
      - id
      - name
      - archived
      - createdAt
      properties:
        id:
          type: string
          description: Server-assigned unique identifier (UUID/cuid).
        name:
          type: string
          minLength: 1
          description: Human-readable subject name.
        color:
          type: string
          description: Optional display color (e.g. hex token) for the subject.
        icon:
          type: string
          description: Optional icon identifier for the subject.
        description:
          type: string
          description: Optional longer description of the subject.
        archived:
          type: boolean
          default: false
          description: Whether the subject is archived (hidden from active lists).
        createdAt:
          type: string
          format: date-time
          description: RFC 3339 creation timestamp.
    CreateSubject:
      type: object
      description: Request body to create a Subject; server assigns id/createdAt and defaults archived=false.
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          description: Human-readable subject name.
        color:
          type: string
          description: Optional display color (e.g. hex token) for the subject.
        icon:
          type: string
          description: Optional icon identifier for the subject.
        description:
          type: string
          description: Optional longer description of the subject.
    PageOfSubject:
      type: object
      description: A single page of Subject results (Page<Subject> per the common pagination pattern).
      required:
      - items
      - total
      - page
      - pageSize
      properties:
        items:
          type: array
          description: The Subjects in this page.
          items:
            $ref: '#/components/schemas/Subject'
        total:
          type: integer
          description: Total number of Subjects matching the query across all pages.
        page:
          type: integer
          description: 1-based page number of this result.
        pageSize:
          type: integer
          description: Maximum number of items per page.
...
    StartAttemptResult:
      type: object
      description: The result of starting an attempt — the attempt plus its delivered questions.
      additionalProperties: false
      required:
      - attempt
      - questions
      properties:
        attempt:
          $ref: '#/components/schemas/ExamAttempt'
        questions:
          type: array
          description: The delivered questions, in delivery order, without the answer key.
          items:
            $ref: '#/components/schemas/DeliveredQuestion'
...
    StudentSignIn:
      type: object
      description: Body to sign a student in for an exam session.
      additionalProperties: false
      required:
      - studentId
      properties:
        studentId:
          type: string
          description: The student signing in.
        pin:
          type: string
          description: Optional PIN gating the student sign-in.
    AuthResult:
      type: object
      description: The response of a WebAuthn ceremony step. During the begin step it carries opaque ceremony
        options; on a successful finish it carries the established Session. Both fields are optional;
        exactly one is present per step.
      additionalProperties: false
      properties:
        options:
          $ref: '#/components/schemas/WebAuthnCeremony'
        session:
          $ref: '#/components/schemas/Session'
...